third_party/blink/web_tests/wpt_internal/trust-tokens/trust-token-api-e2e-iframe.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Tests trust token issuance and redemption</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<body>
    <script>
        'use strict';

        promise_test(function () {
            return fetch('/wpt_internal/trust-tokens/resources/trust_token_issuance.py', {
                privateToken: { type: 'private-state-token', version: 1, operation: 'token-request' }
            }).then(function (response) {
                assert_equals(response.status, 200);
                assert_equals(response.headers.get('Sec-Private-State-Token'), null);
            })
        }, 'Token issuance succeeds.');

        promise_test(function () {
            return fetch('/wpt_internal/trust-tokens/resources/trust_token_redemption.py', {
                privateToken: { type: 'private-state-token', version: 1, operation: 'token-redemption' }
            }).then(async function (response) {
                assert_equals(response.status, 200);
                assert_equals(response.headers.get('Sec-Private-State-Token'), null);
                const redemptionResponse = await response.text()
                assert_equals(redemptionResponse, 'dummy redemption record');
            })
        }, 'Token redemption succeeds.');

        promise_test(() => {
            const host_info = get_host_info();
            const frame = document.createElement('iframe');
            frame.src = '/wpt_internal/trust-tokens/resources/trust_token_send_redemption_record.py';
            frame.privateToken = JSON.stringify({ type: 'private-state-token', version: 1, operation: 'send-redemption-record', issuers: [host_info.ORIGIN] });
            document.body.appendChild(frame);

            return new Promise(resolve => {
                frame.addEventListener("load", resolve);
            }).then(event => {
                const redemptionRecord = frame.contentWindow.document.body.innerText;
                assert_true(redemptionRecord.includes('redemption-record="dummy redemption record"'));
            });
        }, 'Trust token RR succeeds using iframes.');
    </script>
</body>
